+++ common/eth2_network_config/Cargo.toml
@@ -17,7 +17,6 @@ tempfile = "3.1.0"
 serde_yaml = "0.8.13"
 serde_json = "1.0.58"
 types = { path = "../../consensus/types"}
-kzg = { path = "../../crypto/kzg" }
 ethereum_ssz = "0.5.0"
 eth2_config = { path = "../eth2_config"}
 discv5 = "0.3.1"
