+++ lcli/src/new_testnet.rs
@@ -1,7 +1,7 @@
 use account_utils::eth2_keystore::keypair_from_secret;
 use clap::ArgMatches;
 use clap_utils::{parse_optional, parse_required, parse_ssz_optional};
-use eth2_network_config::{get_trusted_setup, Eth2NetworkConfig};
+use eth2_network_config::{Eth2NetworkConfig};
 use eth2_wallet::bip39::Seed;
 use eth2_wallet::bip39::{Language, Mnemonic};
 use eth2_wallet::{recover_validator_secret_from_mnemonic, KeyType};
@@ -196,25 +196,12 @@ pub fn run<T: EthSpec>(testnet_dir_path: PathBuf, matches: &ArgMatches) -> Resul
         None
     };
 
-    let kzg_trusted_setup = if let Some(epoch) = spec.deneb_fork_epoch {
-        // Only load the trusted setup if the deneb fork epoch is set
-        if epoch != Epoch::max_value() {
-            let trusted_setup: TrustedSetup =
-                serde_json::from_reader(get_trusted_setup::<T::Kzg>())
-                    .map_err(|e| format!("Unable to read trusted setup file: {}", e))?;
-            Some(trusted_setup)
-        } else {
-            None
-        }
-    } else {
-        None
-    };
+    
     let testnet = Eth2NetworkConfig {
         deposit_contract_deploy_block,
         boot_enr: Some(vec![]),
         genesis_state_bytes,
         config: Config::from_chain_spec::<T>(&spec),
-        kzg_trusted_setup,
     };
 
     testnet.write_to_file(testnet_dir_path, overwrite_files)
